requirements.txt
Flask
flask-cors
selenium
webdriver-manager
beautifulsoup4
requests
numpy
pandas
Pillow
tensorflow
scikit-learn